pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trustyai"
version = "0.1.0"
description = "An LLM wrapper that estimates the trustworthiness of the LLM's answers"
authors = [{name = "Khanh Do", email = "dokhanh@gmail.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "google-genai>=1.26",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.5",
]

[project.urls]
Homepage = "https://github.com/kndo/trustyai"
Repository = "https://github.com/kndo/trustyai"

[tool.ruff]
line-length = 120
